+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/service/impl/DefaultIdmRoleRequestService.java
@@ -421,7 +421,7 @@ public class DefaultIdmRoleRequestService
 		}
 
 		if (requestDto != null && requestDto.getWfProcessId() != null) {
-			WorkflowProcessInstanceDto processDto = workflowProcessInstanceService.get(requestDto.getWfProcessId());
+			WorkflowProcessInstanceDto processDto = workflowProcessInstanceService.get(requestDto.getWfProcessId(), false);
 			// TODO: create trimmed variant in workflow process instance service
 			if (processDto != null) {
 				processDto.setProcessVariables(null);