+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/workflow/model/dto/WorkflowProcessInstanceDto.java
@@ -5,10 +5,19 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.springframework.hateoas.core.Relation;
 import org.springframework.util.Assert;
 
 import eu.bcvsolutions.idm.core.api.dto.BaseDto;
 
+
+/**
+ * 
+ * @author svandav
+ *
+ */
+//TODO: Use in the since version 8.x.x @Relation(collectionRelation = "workflowProcessInstances")
+@Relation(collectionRelation = "resources")
 public class WorkflowProcessInstanceDto implements BaseDto{
 
 	private static final long serialVersionUID = -719948059760251573L;