+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/workflow/model/dto/WorkflowTaskInstanceDto.java
@@ -1,21 +1,29 @@
 package eu.bcvsolutions.idm.core.workflow.model.dto;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.springframework.hateoas.core.Relation;
+import org.springframework.util.Assert;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import eu.bcvsolutions.idm.core.api.dto.BaseDto;
 
 /**
  * 
  * @author svandav
  *
  */
+//TODO: Use in the since version 8.x.x @Relation(collectionRelation = "workflowTaskInstances")
+@Relation(collectionRelation = "resources")
+public class WorkflowTaskInstanceDto implements BaseDto {
 
-public class WorkflowTaskInstanceDto {
-
+	private static final long serialVersionUID = 1L;
 	private String id;
 	@JsonProperty(value = "taskName")
 	private String name;
@@ -162,4 +170,11 @@ public class WorkflowTaskInstanceDto {
 		this.formKey = formKey;
 	}
 
+	@Override
+	public void setId(Serializable id) {
+		if (id != null) {
+			Assert.isInstanceOf(String.class, id, "WorkflowTaskInstanceDto supports only String identifier.");
+		}
+		this.id = (String) id;
+	}
 }