+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/workflow/service/impl/DefaultWorkflowHistoricProcessInstanceService.java
@@ -1,6 +1,7 @@
 package eu.bcvsolutions.idm.core.workflow.service.impl;
 
 import java.io.InputStream;
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
@@ -27,10 +28,21 @@ import org.activiti.engine.impl.pvm.process.ActivityImpl;
 import org.activiti.engine.runtime.ProcessInstance;
 import org.activiti.image.ProcessDiagramGenerator;
 import org.activiti.image.impl.DefaultProcessDiagramGenerator;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.domain.Sort.Order;
 import org.springframework.stereotype.Service;
+import org.springframework.util.Assert;
 
 import eu.bcvsolutions.idm.core.api.rest.domain.ResourcesWrapper;
+import eu.bcvsolutions.idm.core.rest.AbstractBaseDtoService;
+import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
 import eu.bcvsolutions.idm.core.security.api.service.SecurityService;
 import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowFilterDto;
 import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowHistoricProcessInstanceDto;
@@ -44,7 +56,7 @@ import eu.bcvsolutions.idm.core.workflow.service.WorkflowProcessDefinitionServic
  *
  */
 @Service
-public class DefaultWorkflowHistoricProcessInstanceService implements WorkflowHistoricProcessInstanceService {
+public class DefaultWorkflowHistoricProcessInstanceService extends AbstractBaseDtoService<WorkflowHistoricProcessInstanceDto, WorkflowFilterDto> implements WorkflowHistoricProcessInstanceService {
 
 	private static final String DEFINITION_ID_DELIMITER = ":";
 
@@ -63,15 +75,9 @@ public class DefaultWorkflowHistoricProcessInstanceService implements WorkflowHi
 	@Autowired
 	private WorkflowProcessDefinitionService definitionService;
 
-	/**
-	 * Search process history. Process variables will be included only for get
-	 * specific process history. It means filter.processInstanceId is filled.
-	 * 
-	 * @param filter
-	 * @return
-	 */
 	@Override
-	public ResourcesWrapper<WorkflowHistoricProcessInstanceDto> search(WorkflowFilterDto filter) {
+	public Page<WorkflowHistoricProcessInstanceDto> find(WorkflowFilterDto filter, Pageable pageable,
+			BasePermission... permission) {
 		String processDefinitionId = filter.getProcessDefinitionId();
 		String processInstanceId = filter.getProcessInstanceId();
 
@@ -108,44 +114,102 @@ public class DefaultWorkflowHistoricProcessInstanceService implements WorkflowHi
 		
 		// check security ... only involved user or applicant can work with
 		// historic process instance ... admin can see all historic processes every time
+		// TODO: refactor and use username/id from filter
 		if(!securityService.isAdmin()) {
 			// Applicant and Implementer is added to involved user after process
 			// (subprocess) started. This modification allow not use OR clause.
 			query.involvedUser(securityService.getCurrentId().toString());
 		}
 
-		if (WorkflowHistoricProcessInstanceService.SORT_BY_START_TIME.equals(filter.getSortByFields())) {
+		String fieldForSort = null;
+		boolean ascSort = false;
+		boolean descSort = false;
+		if (pageable != null) {
+			Sort sort = pageable.getSort();
+			if (sort != null) {
+				for (Order order : sort) {
+					if (!StringUtils.isEmpty(order.getProperty())) {
+						// TODO: now is implemented only one property sort 
+						fieldForSort = order.getProperty();
+						if (order.getDirection() == Direction.ASC) {
+							ascSort = true;
+						} else if (order.getDirection() == Direction.DESC) {
+							descSort = true;
+						}
+						break;
+					}
+					
+				}
+			}
+		}
+
+		if (WorkflowHistoricProcessInstanceService.SORT_BY_START_TIME.equals(fieldForSort)) {
 			query.orderByProcessInstanceStartTime();
-		} else if (WorkflowHistoricProcessInstanceService.SORT_BY_END_TIME.equals(filter.getSortByFields())) {
+		} else if (WorkflowHistoricProcessInstanceService.SORT_BY_END_TIME.equals(fieldForSort)) {
 			query.orderByProcessInstanceEndTime();
 		} else {
 			query.orderByProcessDefinitionId();
-		}
-		if (filter.isSortAsc()) {
+			// there must be default order
 			query.asc();
 		}
-		if (filter.isSortDesc()) {
+		if (ascSort) {
+			query.asc();
+		}
+		if (descSort) {
 			query.desc();
 		}
 		long count = query.count();
-		List<HistoricProcessInstance> processInstances = query.listPage((filter.getPageNumber()) * filter.getPageSize(),
-				filter.getPageSize());
+		
+		// it's possible that pageable is null
+		List<HistoricProcessInstance> processInstances = null;
+		if (pageable == null) {
+			processInstances = query.list();
+		} else {
+			processInstances = query.listPage((pageable.getPageNumber()) * pageable.getPageSize(), pageable.getPageSize());
+		}
+		
 		List<WorkflowHistoricProcessInstanceDto> dtos = new ArrayList<>();
-
 		if (processInstances != null) {
 			for (HistoricProcessInstance instance : processInstances) {
 				dtos.add(toResource(instance, trimmed));
 			}
 		}
-		double totalPageDouble = ((double) count / filter.getPageSize());
-		double totlaPageFlorred = Math.floor(totalPageDouble);
-		long totalPage = 0;
-		if (totalPageDouble > totlaPageFlorred) {
-			totalPage = (long) (totlaPageFlorred + 1);
+		
+		return new PageImpl<WorkflowHistoricProcessInstanceDto>(dtos, pageable, count);
+	}
+	
+	/**
+	 * Search process history. Process variables will be included only for get
+	 * specific process history. It means filter.processInstanceId is filled.
+	 * 
+	 * @param filter
+	 * @return
+	 */
+	@Override
+	public ResourcesWrapper<WorkflowHistoricProcessInstanceDto> search(WorkflowFilterDto filter) {
+		Pageable pageable = null;
+		// get pageable setting from filter - backward compatibility
+		if (StringUtils.isNotEmpty(filter.getSortByFields())) {
+			Sort sort = null;
+			if (filter.isSortAsc()) {
+				sort = new Sort(Direction.ASC, filter.getSortByFields());	
+			} else {
+				sort = new Sort(Direction.DESC, filter.getSortByFields());
+			}
+			pageable = new PageRequest(filter.getPageNumber(), filter.getPageSize(), sort);
+		} else {
+			pageable = new PageRequest(filter.getPageNumber(), filter.getPageSize());
 		}
-
-		return new ResourcesWrapper<>(dtos, count, totalPage,
+		Page<WorkflowHistoricProcessInstanceDto> page = this.find(filter, pageable);
+		
+		return new ResourcesWrapper<>(page.getContent(), page.getTotalElements(), page.getTotalPages(),
 				filter.getPageNumber(), filter.getPageSize());
+	}
+	
+	@Override
+	public WorkflowHistoricProcessInstanceDto get(Serializable id, BasePermission... permission) {
+		Assert.notNull(id);
+		return this.get(String.valueOf(id));
 	}
 
 	@Override